requirements.txt
requests>=2.28
scikit-learn>=1.3
numpy>=1.24
python-dotenv>=1.0
